remove stubs, comment out unimplemented, fix build errors
- Remove stubs.c entirely - only export functions we actually implement
- Comment out all unimplemented function declarations in msi_interop.h
  so the header itself serves as documentation of what's available
- Fix DWORD* type consistency: string_utils uses DWORD* (matching
  LPDWORD) instead of unsigned long* for cross-platform compatibility
- Fix libmsi-enums.h not found: add generated header as dependency
  via libmsi_enums[1] (header only, not .c to avoid duplicate symbols)

+++ msi-interop/string_utils.c
@@ -52,7 +52,7 @@ utf8_to_utf16(const char *utf8, size_t *out_len)
 }
 
 unsigned int
-copy_string_to_bufA(const char *str, char *buf, unsigned long *pcch)
+copy_string_to_bufA(const char *str, char *buf, DWORD *pcch)
 {
     if (str == NULL)
         str = "";
@@ -103,7 +103,7 @@ copy_string_to_bufA(const char *str, char *buf, unsigned long *pcch)
 }
 
 unsigned int
-copy_string_to_bufW(const char *str, WCHAR *buf, unsigned long *pcch)
+copy_string_to_bufW(const char *str, WCHAR *buf, DWORD *pcch)
 {
     if (str == NULL)
         str = "";
